requirements.txt
Django==5.2.2
djangorestframework==3.16.0